give only superuser access to edit and delete products

+++ products/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from .models import Product, Category
 from .forms import ProductForm
@@ -67,8 +68,12 @@ def product_detail(request, product_id):
 
     return render(request, 'products/product_detail.html', context)
 
-
+@login_required
 def add_product(request):
+     if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners have access.')
+        return redirect(reverse('core'))
+
      if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -88,8 +93,11 @@ def add_product(request):
       return render(request, template, context)
 
 
-
+@login_required
 def edit_product(request, product_id):
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners have access.')
+        return redirect(reverse('core'))
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
          form = ProductForm(request.POST, request.FILES, instance=product)
@@ -111,8 +119,11 @@ def edit_product(request, product_id):
 
     return render(request, template, context)
 
-
+@login_required
 def delete_product(request, product_id):
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners have access.')
+        return redirect(reverse('core'))
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Product deleted!')
